+++ db/schema.rb
@@ -129,6 +129,7 @@ ActiveRecord::Schema.define(version: 20161031153002) do
     t.string   "current_address_state"
     t.string   "current_address_country"
     t.string   "current_address_postal"
+    t.integer  "user_id"
   end
 
   create_table "billing_log_reports", force: :cascade do |t|