+++ CHANGELOG.md
@@ -6,6 +6,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 ### Added
 - Invitations can now be reordered through drag and drop. This also enforces
   rules automatically for which invitations can be reordered and into which groups
+- A user can connect their ORCID account with their author profile via the Authors cards on a manuscript
 ### Changed
 ### Deprecated
 ### Removed